reprod/requirements.txt
pyyaml==6.0.2
openai==1.101.0
matplotlib==3.10.5
networkx===3.5
